[Fix] Fix e-mail submission (#7)
* Switch order of input name and type

* Add hidden form to help netlify know about fields

* Added honeypot to form

* Separated email subscribe form to own component and did additional cleanup
